These tables were removed * Classes * Attendance

+++ supabase/migrations/20260716232000_add_class_action_windows.sql
@@ -1,10 +1,8 @@
--- Per-class join/exit windows, replacing the hardcoded 10-minute buffer.
--- join_window_minutes: how long before start_time joining opens.
--- exit_window_minutes: how long after end_time exiting stays open.
+-- Add per-class join/exit windows for attendance tracking.
 
 alter table public.classes
-  add column join_window_minutes smallint not null default 10,
-  add column exit_window_minutes smallint not null default 10;
+  add column join_window_minutes integer not null default 10,
+  add column exit_window_minutes integer not null default 10;
 
 alter table public.classes
   add constraint classes_join_window_valid check (join_window_minutes between 0 and 120),
